Fixed hyprland, hyprpaper, and neovim configs, fixed prayer.sh
Updated hyprland.conf due to update changing how shadows are configured.

Added empty hyprpaper.conf due to update causing hyprpaper daemon to not run if it does not find a configuration file.

Fixed deprecated name of JS/TS LSP from tsserver to ts_ls in nvim config.

Fixed deprecated neoscroll function call.

Fixed incorrect handling of Isha/Midnight/Last Third times when midnight time is not after 00:00 (ex.: when midnight is at 23:20). Previously, midnight time would be saved as 00:-40 instead of 23:20. Now, when the current time is before Fajr, the logic is handled manually with multiple if statements rather than in the main while loop. As a result, the tooltip for the Last Third (Fajr in...) is also fixed (it would not appear previously). Additionally, the order in which the prayer times are saved has been changed. Midnight and Last third times are now at the end of the file rather than the top. The previous order was necessary to handle midnight/last third times but did not support midnight times before 00:00. Also, prayer times from mawaqit.net are now fetched by day and month rather than "today's times". This would cause an issue where the incorrect prayer times are saved when the date was changed by the script (ex.: yesterday's date).